Add Service interfaces and CvService unit test.
