+++ build.gradle
@@ -31,4 +31,5 @@ dependencies {
     compile group: 'org.javassist', name: 'javassist', version: '3.24.1-GA'
     testCompile group: 'org.spockframework', name: 'spock-core', version: '1.3-groovy-2.5'
     testCompile group: 'org.codehaus.groovy', name: 'groovy-test', version: '2.5.6'
+    testCompile group: 'uk.co.jemos.podam', name: 'podam', version: '7.2.2.RELEASE'
 }
