+++ src/main/java/com/billennium/bcvsaver/dto/ProjectDto.java
@@ -2,11 +2,11 @@ package com.billennium.bcvsaver.dto;
 
 import lombok.Data;
 
-import java.util.Set;
-
 @Data
 public class ProjectDto {
     private String client;
     private String position;
+    // TODO
+    // project has many technologies
     //private Set<String> technologies;
 }
